working on getting all the components for dr
